shopFinderFirebase/requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
google-auth==1.6.2
requests==2.21.0
Flask-Session
google-cloud-storage==1.13.2
firebase-admin
pillow
google-cloud-vision